feat: refactor authentication and role management structure; update imports and interfaces for clarity and consistency
