+++ team-2-group-project/.env.example
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_API_LINK=http://localhost:5000
+NEXT_PUBLIC_API_LINK="http://localhost:43313"
