+++ team-2-group-project/app/page.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 import AppSpinner from "@/components/spinner/AppSpinner";
 import { useAuthState } from "@/providers/authProvider";
-import { selectBestAuthenticatedRoute } from "@/utils/roles";
+import { selectBestAuthenticatedRoute } from "@/utils/auth/roles";
 
 export default function EntryPage() {
   const router = useRouter();
