+++ team-2-group-project/components/auth/AuthFooterLink.tsx
@@ -3,14 +3,9 @@
 import Link from "next/link";
 import { Typography } from "antd";
 import { useStyles } from "@/app/(auth)/style/style";
+import { IAuthFooterLinkProps } from "@/interfaces/auth/authProps";
 
 const { Text } = Typography;
-
-interface IAuthFooterLinkProps {
-  question: string;
-  label: string;
-  href: string;
-}
 
 const AuthFooterLink = ({
   question,
