+++ team-2-group-project/constants/auth/roles.ts
@@ -1,0 +1,10 @@
+export const PERMISSIONS = {
+  users: "Pages.Users",
+  roles: "Pages.Roles",
+  tenants: "Pages.Tenants",
+} as const;
+
+export const STATIC_ROLES = {
+  hostAdmin: "Host.Admin",
+  tenantAdmin: "Tenants.Admin",
+} as const;
