+++ team-2-group-project/constants/auth/tenantService.ts
@@ -1,0 +1,4 @@
+export const TENANT_COOKIE_KEY = "Abp.TenantId";
+export const TENANT_ID_STORAGE_KEY = "abp_tenant_id";
+export const TENANCY_NAME_STORAGE_KEY = "abp_tenancy_name";
+export const TENANCY_NAME_QUERY_KEY = "abp_tenancy_name";
