+++ team-2-group-project/enums/auth/authService.ts
@@ -1,0 +1,3 @@
+export const TOKEN_STORAGE_KEY = "access_token";
+export const REMEMBERED_TOKEN_STORAGE_KEY = "remembered_access_token";
+export const ENCRYPTED_TOKEN_COOKIE_KEY = "enc_auth_token";
