+++ team-2-group-project/enums/auth/tenantService.ts
@@ -1,0 +1,5 @@
+export enum TenantAvailabilityState {
+  Available = 1,
+  InActive = 2,
+  NotFound = 3,
+}
