+++ team-2-group-project/interfaces/abp.ts
@@ -1,0 +1,11 @@
+export interface IAbpResponse<T> {
+  result?: T | null;
+  targetUrl?: string | null;
+  success?: boolean;
+  error?: {
+    message?: string;
+    details?: string;
+  } | null;
+  unAuthorizedRequest?: boolean;
+  __abp?: boolean;
+}
