+++ team-2-group-project/interfaces/auth/roles.ts
@@ -1,0 +1,4 @@
+export interface IRouteSelectorUser {
+  permissions?: string[] | null;
+  roles?: string[] | null;
+}
