+++ team-2-group-project/interfaces/auth/tenantService.ts
@@ -1,0 +1,11 @@
+import { TenantAvailabilityState } from "@/enums/auth/tenantService";
+
+export interface ITenantContext {
+  tenantId?: number | null;
+  tenancyName?: string | null;
+}
+
+export interface IIsTenantAvailableResponse {
+  state: TenantAvailabilityState;
+  tenantId?: number | null;
+}
