+++ team-2-group-project/interfaces/spinnerProps.ts
@@ -1,0 +1,3 @@
+export interface IAppSpinnerProps {
+  label?: string;
+}
