+++ team-2-group-project/utils/auth/adminService.ts
@@ -1,62 +1,21 @@
 import { axiosInstance } from "@/utils/axiosInstance";
-import { IAbpResponse, unwrapAbpResponse } from "@/utils/abp";
+import { unwrapAbpResponse } from "@/utils/abp";
+import { IAbpResponse } from "@/interfaces/abp";
+import { IChangePasswordRequest, IPagedResult, IRoleListItem, ITenantListItem, IUserListItem } from "@/interfaces/auth/adminService";
 
-export interface IPagedResult<T> {
-  items?: T[] | null;
-  totalCount?: number;
-}
 
-export interface IUserListItem {
-  id: number;
-  userName: string;
-  name: string;
-  surname: string;
-  fullName?: string | null;
-  emailAddress: string;
-  isActive: boolean;
-  creationTime?: string | null;
-  lastLoginTime?: string | null;
-  roleNames?: string[] | null;
-}
-
-export interface IRoleListItem {
-  id: number;
-  name: string;
-  displayName: string;
-  normalizedName?: string | null;
-  description?: string | null;
-  grantedPermissions?: string[] | null;
-  isStatic?: boolean;
-  isDefault?: boolean;
-  creationTime?: string | null;
-}
-
-export interface ITenantListItem {
-  id: number;
-  tenancyName: string;
-  name: string;
-  isActive: boolean;
-}
-
-export interface IChangePasswordRequest {
-  currentPassword: string;
-  newPassword: string;
-}
 
 export const getUsers = async () => {
   const response = await axiosInstance().get<
     IAbpResponse<IPagedResult<IUserListItem>>
-  >(
-    "/api/services/app/User/GetAll",
-    {
-      params: {
-        keyword: "",
-        isActive: undefined,
-        skipCount: 0,
-        maxResultCount: 100,
-      },
+  >("/api/services/app/User/GetAll", {
+    params: {
+      keyword: "",
+      isActive: undefined,
+      skipCount: 0,
+      maxResultCount: 100,
     },
-  );
+  });
 
   return unwrapAbpResponse(response.data);
 };
@@ -64,16 +23,13 @@ export const getUsers = async () => {
 export const getRoles = async () => {
   const response = await axiosInstance().get<
     IAbpResponse<IPagedResult<IRoleListItem>>
-  >(
-    "/api/services/app/Role/GetAll",
-    {
-      params: {
-        keyword: "",
-        skipCount: 0,
-        maxResultCount: 100,
-      },
+  >("/api/services/app/Role/GetAll", {
+    params: {
+      keyword: "",
+      skipCount: 0,
+      maxResultCount: 100,
     },
-  );
+  });
 
   return unwrapAbpResponse(response.data);
 };
@@ -81,17 +37,14 @@ export const getRoles = async () => {
 export const getTenants = async () => {
   const response = await axiosInstance().get<
     IAbpResponse<IPagedResult<ITenantListItem>>
-  >(
-    "/api/services/app/Tenant/GetAll",
-    {
-      params: {
-        keyword: "",
-        isActive: undefined,
-        skipCount: 0,
-        maxResultCount: 100,
-      },
+  >("/api/services/app/Tenant/GetAll", {
+    params: {
+      keyword: "",
+      isActive: undefined,
+      skipCount: 0,
+      maxResultCount: 100,
     },
-  );
+  });
 
   return unwrapAbpResponse(response.data);
 };
